@@ -1,5 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
+import Ticket from "../Ticket/Ticket";
 
 const LaneWrapper = styled.div`
   list-style: none;
@@ -22,10 +23,22 @@ const Title = styled.h2`
   border-bottom: 1px solid darkGray;
 `;
 
-const Lane = ({ title }) => (
-  <LaneWrapper>
-    <Title>{title}</Title>
-  </LaneWrapper>
+const TicketsWrapper = styled.div`
+    padding: 5%;
+`;
+
+const Alert = styled.div`
+    text-align: center;
+`;
+
+const Lane = ({title, loading, error, tickets, onDragStart, onDragOver, onDrop, laneId}) => (
+	<LaneWrapper onDragOver={onDragOver} onDrop={e => onDrop(e, laneId)}>
+		<Title>{title}</Title>
+		{ (loading || error ) && <Alert>{ loading ? 'Loading...' : error }</Alert> }
+		<TicketsWrapper>
+			{ tickets.map(ticket => <Ticket key={ticket.id} ticket={ticket} onDragStart={onDragStart}/>) }
+		</TicketsWrapper>
+	</LaneWrapper>
 );
 
 export default Lane;
