@@ -1,7 +1,8 @@
-import React, { Component } from 'react';
-import styled, { createGlobalStyle } from 'styled-components';
+import React, {Component} from 'react';
+import styled, {createGlobalStyle} from 'styled-components';
 import Board from './Board';
 import Header from '../components/Header/Header';
+import Tickets from "../components/Tickets/Tickets";
 
 const GlobalStyle = createGlobalStyle`
   body {
@@ -20,17 +21,25 @@ const AppWrapper = styled.div`
 `;
 
 class App extends Component {
-  render() {
-    return (
-      <>
-        <GlobalStyle />
-        <AppWrapper>
-          <Header />
-          <Board />
-        </AppWrapper>
-      </>
-    );
-  }
+	render() {
+		const lanes = [
+			{id: 1, title: 'To Do'},
+			{id: 2, title: 'In Progress'},
+			{id: 3, title: 'Review'},
+			{id: 4, title: 'Done'},
+		];
+
+		return (
+			<>
+				<GlobalStyle/>
+				<AppWrapper>
+					<Header/>
+					<Board lanes={lanes} dataSource="/assets/data.json"/>
+					<Tickets dataSource="/assets/data.json"/>
+				</AppWrapper>
+			</>
+		);
+	}
 }
 
 export default App;
